fix(extrinsics): define CLI structs in cargo contract
