+++ crates/cargo-contract/Cargo.toml
@@ -35,6 +35,7 @@ url = { version = "2.4.1", features = ["serde"] }
 # dependencies for extrinsics (deploying and calling a contract)
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 subxt = "0.31.0"
+sp-core = "21.0.0"
 sp-weights = "20.0.0"
 hex = "0.4.3"
