+++ crates/cargo-contract/src/cmd/call.rs
@@ -1,0 +1,205 @@
+use crate::ErrorVariant;
+
+use contract_build::util::DEFAULT_KEY_COL_WIDTH;
+use std::fmt::Debug;
+use tokio::runtime::Runtime;
+
+use anyhow::{
+    anyhow,
+    Context,
+    Result,
+};
+use contract_build::name_value_println;
+use contract_extrinsics::{
+    display_contract_exec_result,
+    display_contract_exec_result_debug,
+    display_dry_run_result_warning,
+    print_dry_running_status,
+    print_gas_required_success,
+    prompt_confirm_tx,
+    BalanceVariant,
+    CallCommandBuilder,
+    CallDryRunResult,
+    CallExec,
+    DefaultConfig,
+    ExtrinsicOpts,
+    StorageDeposit,
+    TokenMetadata,
+    MAX_KEY_COL_WIDTH,
+};
+use sp_weights::Weight;
+use subxt::Config;
+
+#[derive(Debug, clap::Args)]
+#[clap(name = "call", about = "Call a contract")]
+pub struct CallCommand {
+    /// The address of the the contract to call.
+    #[clap(name = "contract", long, env = "CONTRACT")]
+    contract: <DefaultConfig as Config>::AccountId,
+    /// The name of the contract message to call.
+    #[clap(long, short)]
+    message: String,
+    /// The arguments of the contract message to call.
+    #[clap(long, num_args = 0..)]
+    args: Vec<String>,
+    #[clap(flatten)]
+    extrinsic_opts: ExtrinsicOpts,
+    /// Maximum amount of gas (execution time) to be used for this command.
+    /// If not specified will perform a dry-run to estimate the gas consumed for the
+    /// call.
+    #[clap(name = "gas", long)]
+    gas_limit: Option<u64>,
+    /// Maximum proof size for this call.
+    /// If not specified will perform a dry-run to estimate the proof size required for
+    /// the call.
+    #[clap(long)]
+    proof_size: Option<u64>,
+    /// The value to be transferred as part of the call.
+    #[clap(name = "value", long, default_value = "0")]
+    value: BalanceVariant,
+    /// Export the call output in JSON format.
+    #[clap(long, conflicts_with = "verbose")]
+    output_json: bool,
+}
+
+impl CallCommand {
+    /// Returns whether to export the call output in JSON format.
+    pub fn output_json(&self) -> bool {
+        self.output_json
+    }
+}
+pub fn handle_call(call_command: &CallCommand) -> Result<(), ErrorVariant> {
+    Runtime::new()?.block_on(async {
+        let call_exec = CallCommandBuilder::default()
+            .contract(call_command.contract.clone())
+            .message(call_command.message.clone())
+            .args(call_command.args.clone())
+            .extrinsic_opts(call_command.extrinsic_opts.clone())
+            .gas_limit(call_command.gas_limit)
+            .proof_size(call_command.proof_size)
+            .value(call_command.value.clone())
+            .output_json(call_command.output_json)
+            .done()
+            .await;
+
+        if !call_exec.opts().execute() {
+            let result = call_exec.call_dry_run().await?;
+            match result.result {
+                Ok(ref ret_val) => {
+                    let value = call_exec
+                        .transcoder()
+                        .decode_message_return(
+                            call_exec.message(),
+                            &mut &ret_val.data[..],
+                        )
+                        .context(format!(
+                            "Failed to decode return value {:?}",
+                            &ret_val
+                        ))?;
+                    let dry_run_result = CallDryRunResult {
+                        reverted: ret_val.did_revert(),
+                        data: value,
+                        gas_consumed: result.gas_consumed,
+                        gas_required: result.gas_required,
+                        storage_deposit: StorageDeposit::from(&result.storage_deposit),
+                    };
+                    if call_exec.output_json() {
+                        println!("{}", dry_run_result.to_json()?);
+                    } else {
+                        dry_run_result.print();
+                        display_contract_exec_result_debug::<_, DEFAULT_KEY_COL_WIDTH>(
+                            &result,
+                        )?;
+                        display_dry_run_result_warning("message");
+                    };
+                }
+                Err(ref err) => {
+                    let metadata = call_exec.client().metadata();
+                    let object = ErrorVariant::from_dispatch_error(err, &metadata)?;
+                    if call_exec.output_json() {
+                        return Err(object)
+                    } else {
+                        name_value_println!("Result", object, MAX_KEY_COL_WIDTH);
+                        display_contract_exec_result::<_, MAX_KEY_COL_WIDTH>(&result)?;
+                    }
+                }
+            }
+        } else {
+            let gas_limit = pre_submit_dry_run_gas_estimate_call(&call_exec).await?;
+            if !call_exec.opts().skip_confirm() {
+                prompt_confirm_tx(|| {
+                    name_value_println!(
+                        "Message",
+                        call_exec.message(),
+                        DEFAULT_KEY_COL_WIDTH
+                    );
+                    name_value_println!(
+                        "Args",
+                        call_exec.args().join(" "),
+                        DEFAULT_KEY_COL_WIDTH
+                    );
+                    name_value_println!(
+                        "Gas limit",
+                        gas_limit.to_string(),
+                        DEFAULT_KEY_COL_WIDTH
+                    );
+                })?;
+            }
+            let token_metadata = TokenMetadata::query(call_exec.client()).await?;
+            let display_events = call_exec.call(Some(gas_limit)).await?;
+            let output = if call_exec.output_json() {
+                display_events.to_json()?
+            } else {
+                display_events
+                    .display_events(call_exec.opts().verbosity()?, &token_metadata)?
+            };
+            println!("{output}");
+        }
+        Ok(())
+    })
+}
+
+/// A helper function to estimate the gas required for a contract call.
+async fn pre_submit_dry_run_gas_estimate_call(call_exec: &CallExec) -> Result<Weight> {
+    if call_exec.opts().skip_dry_run() {
+        return match (call_exec.gas_limit(), call_exec.proof_size()) {
+            (Some(ref_time), Some(proof_size)) => Ok(Weight::from_parts(ref_time, proof_size)),
+            _ => {
+                Err(anyhow!(
+                "Weight args `--gas` and `--proof-size` required if `--skip-dry-run` specified"
+            ))
+            }
+        };
+    }
+    if !call_exec.output_json() {
+        print_dry_running_status(call_exec.message());
+    }
+    let call_result = call_exec.call_dry_run().await?;
+    match call_result.result {
+        Ok(_) => {
+            if !call_exec.output_json() {
+                print_gas_required_success(call_result.gas_required);
+            }
+            // use user specified values where provided, otherwise use the estimates
+            let ref_time = call_exec
+                .gas_limit()
+                .unwrap_or_else(|| call_result.gas_required.ref_time());
+            let proof_size = call_exec
+                .proof_size()
+                .unwrap_or_else(|| call_result.gas_required.proof_size());
+            Ok(Weight::from_parts(ref_time, proof_size))
+        }
+        Err(ref err) => {
+            let object =
+                ErrorVariant::from_dispatch_error(err, &call_exec.client().metadata())?;
+            if call_exec.output_json() {
+                Err(anyhow!("{}", serde_json::to_string_pretty(&object)?))
+            } else {
+                name_value_println!("Result", object, MAX_KEY_COL_WIDTH);
+                display_contract_exec_result::<_, MAX_KEY_COL_WIDTH>(&call_result)?;
+
+                Err(anyhow!("Pre-submission dry-run failed. Use --skip-dry-run to skip this step."))
+            }
+        }
+    }
+}
